Remove utils class. Simplify biometric prompt builder
